OmniLuck_Backend_Python/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
attrs==25.4.0
CacheControl==0.14.3
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.3
exceptiongroup==1.3.1
fastapi==0.115.0
filelock==3.19.1
firebase-admin==6.6.0
frozenlist==1.8.0
fsspec==2025.10.0
google-ai-generativelanguage==0.6.10
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.7.0
google-crc32c==1.7.1
google-generativeai==0.8.3
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.62.3
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
networkx==3.2.1
numpy==1.24.4
packaging==25.0
pandas==2.0.3
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.10.0
pydantic-settings==2.6.1
pydantic_core==2.27.0
PyJWT==2.10.1
pyparsing==3.2.5
pyswisseph==2.10.3.2
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
safetensors==0.7.0
scipy==1.13.1
six==1.17.0
starlette==0.38.6
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.15.2
tomli==2.3.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.2
uvicorn==0.32.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
